+++ NEWS.md
@@ -1,6 +1,7 @@
 # mapSpain (development version)
 
 -   Move minimum version of `giscoR` to v0.2.4
+-   Fix typos on `esp_dict_translate()` [\#36](https://github.com/rOpenSpain/mapSpain/issues/36)
 
 # mapSpain 0.2.2